[dy] Add callback helper functions
